Authorize resources using Gates

+++ app/Http/Controllers/TodoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Todo;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class TodoController extends Controller
 {
@@ -60,6 +61,9 @@ class TodoController extends Controller
      */
     public function show(Todo $todo)
     {
+        if (Gate::denies('access-todo', $todo)) {
+            abort(403, 'Unauthorized action.');
+        }
         return view('todo.show', ['todo' => $todo]);
     }
 
@@ -71,6 +75,9 @@ class TodoController extends Controller
      */
     public function edit(Todo $todo)
     {
+        if (Gate::denies('access-todo', $todo)) {
+            abort(403, 'Unauthorized action.');
+        }
         return view('todo.edit', ['todo' => $todo]);
     }
 
@@ -83,6 +90,10 @@ class TodoController extends Controller
      */
     public function update(Request $request, Todo $todo)
     {
+        if (Gate::denies('access-todo', $todo)) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $validatedData = $request->validate([
             'name' => 'required|max:255',
             'description' => '',
@@ -102,6 +113,9 @@ class TodoController extends Controller
      */
     public function destroy(Todo $todo)
     {
+        if (Gate::denies('access-todo', $todo)) {
+            abort(403, 'Unauthorized action.');
+        }
         $todo->delete();
         return redirect('todo');
     }
